Restructure data layer
