+++ app/src/main/java/by/alexandr7035/affinidi_id/data/DicebearAvatarsHelper.kt
@@ -1,5 +1,7 @@
 package by.alexandr7035.affinidi_id.data
 
+import by.alexandr7035.affinidi_id.data.model.profile.DicebearImageType
+
 interface DicebearAvatarsHelper {
     fun getImageUrl(
         // See all the possible types on the site
