--- app/src/main/java/by/alexandr7035/affinidi_id/data/model/SignUpResponse.kt
@@ -1,4 +1,0 @@
-package by.alexandr7035.affinidi_id.data.model
-
-class SignUpResponse {
-}
