--- app/src/main/java/by/alexandr7035/affinidi_id/data/model/UserProfile.kt
@@ -1,6 +1,0 @@
-package by.alexandr7035.affinidi_id.data.model
-
-data class UserProfile(
-    val userName: String,
-    val userDid: String
-)
