+++ app/src/main/java/by/alexandr7035/affinidi_id/data/model/profile/DicebearImageType.kt
@@ -1,4 +1,4 @@
-package by.alexandr7035.affinidi_id.data
+package by.alexandr7035.affinidi_id.data.model.profile
 
 enum class DicebearImageType {
     AVATAR_JDENTICON,
