+++ app/src/main/java/by/alexandr7035/affinidi_id/data/model/profile/UserProfileModel.kt
@@ -1,0 +1,6 @@
+package by.alexandr7035.affinidi_id.data.model.profile
+
+data class UserProfileModel(
+    val userName: String,
+    val userDid: String
+)
