+++ app/src/main/java/by/alexandr7035/affinidi_id/data/model/sign_in/SignInModel.kt
@@ -1,4 +1,4 @@
-package by.alexandr7035.affinidi_id.data.model
+package by.alexandr7035.affinidi_id.data.model.sign_in
 
 import by.alexandr7035.affinidi_id.core.ErrorType
