+++ app/src/main/java/by/alexandr7035/affinidi_id/data/model/sign_up/SignUpMessageParams.kt
@@ -1,4 +1,4 @@
-package by.alexandr7035.affinidi_id.data.model
+package by.alexandr7035.affinidi_id.data.model.sign_up
 
 import com.google.gson.annotations.SerializedName
