+++ app/src/main/java/by/alexandr7035/affinidi_id/data/model/sign_up/SignUpModel.kt
@@ -1,4 +1,4 @@
-package by.alexandr7035.affinidi_id.data.model
+package by.alexandr7035.affinidi_id.data.model.sign_up
 
 import by.alexandr7035.affinidi_id.core.ErrorType
