+++ app/src/main/java/by/alexandr7035/affinidi_id/data/model/sign_up/SignUpResponse.kt
@@ -1,0 +1,4 @@
+package by.alexandr7035.affinidi_id.data.model.sign_up
+
+class SignUpResponse {
+}
